Refactoring: simplify logic in application component. Rename: UserContent -> Body
